% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2.R
\name{converter}
\alias{converter}
\title{Decimal minute to degrees decimal converter}
\usage{
converter(longitude, latitude)
}
\arguments{
\item{longitude}{Data frame column or character vector with longitude coordinates in decimal minutes}

\item{latitude}{Data frame column or character vector with latitude coordinates in decimal minutes}
}
\description{
Takes input as decimal minutes coordinates and returns a data frame formatted in degrees decimal
}
\keyword{convert}
\keyword{coordinates}
